siteblog/siteblog_app/templatetags/menu.py
from django import template
from siteblog_app.models import Category

register = template.Library()

@register.inclusion_tag('siteblog_app/menu_tpl.html')
def show_menu(menu_class='menu'):
    categories = Category.objects.all()
    return {'categories': categories, 'menu_class': menu_class}
