siteblog/requirements.txt
asgiref==3.8.1
django-ckeditor==6.7.2
django-debug-toolbar==5.0.1
django-js-asset==3.0.1
pillow==11.1.0
sqlparse==0.5.3
mysqlclient
python-dotenv
django-analytical